+++ abeja/common/auth.py
@@ -17,8 +17,8 @@ def get_token_from_environment():
     }
 
 
-def get_basic_auth_from_environment():
-    """get abeja platform basic auth id and pass from environment variables
+def get_user_basic_auth_from_environment():
+    """get abeja platform user basic auth id and pass from environment variables
     basic auth id : user_id
     basic auth pass : personal access token of the user
 
@@ -39,6 +39,27 @@ def get_basic_auth_from_environment():
     }
 
 
+def get_datasource_basic_auth_from_environment():
+    """get abeja platform datasource basic auth id and pass from environment variables
+    basic auth id : datasource id
+    basic auth pass : datasource secret of the user
+
+    :return: dict
+    {
+        "datasource_id": ABEJA_PLATFORM_DATASOURCE_ID,
+        "datasource_secret": ABEJA_PLATFORM_DATASOURCE_SECRET
+    }
+    """
+    datasource_id = os.environ.get('ABEJA_PLATFORM_DATASOURCE_ID')
+    datasource_secret = os.environ.get('ABEJA_PLATFORM_DATASOURCE_SECRET')
+    if not datasource_id or not datasource_secret:
+        return None
+    return {
+        "datasource_id": datasource_id,
+        "datasource_secret": datasource_secret
+    }
+
+
 def get_credential():
     """get credential in the following priority
     1. jwt token in environment
@@ -46,5 +67,7 @@ def get_credential():
 
     :return: dict
     """
-    credential = get_token_from_environment() or get_basic_auth_from_environment()
+    credential = get_token_from_environment() \
+        or get_user_basic_auth_from_environment() \
+        or get_datasource_basic_auth_from_environment()
     return credential