+++ abeja/training/job.py
@@ -291,7 +291,7 @@ class Jobs():
         get the total number of training jobs.
 
         Params:
-            - **filter_archived** (bool): **[optional]** whether include archived ones or not. (default is not-filtered)
+            - **filter_archived** (bool): **[optional]** If ``true``, include archived jobs, otherwise exclude archived jobs. (default: ``false``)
             - **offset** (int): **[optional]** paging offset.
             - **limit** (int): **[optional]** paging limit.